+++ R/coefLognorm.R
@@ -67,7 +67,45 @@ getParmsLognormForLowerAndUpperLog <- function(
 #' @return numeric matrix wiht columns `mu` and `sigma`, the parameter of the
 #'   lognormal distribution. Rows correspond to rows of inputs.
 #' @export
-#' @exampleFunction example_getParmsLognormForModeAndUpper
+#' @examples
+#' # example 1: a distribution with mode 1 and upper bound 5
+#' (thetaEst <- getParmsLognormForModeAndUpper(1,5))
+#' mle <- exp(thetaEst[1] - thetaEst[2]^2)
+#' all.equal(mle, 1, check.attributes = FALSE)
+#' 
+#' # plot the distributions
+#' xGrid = seq(0,8, length.out = 81)[-1]
+#' dxEst <- dlnorm(xGrid, meanlog = thetaEst[1], sdlog = thetaEst[2])
+#' plot( dxEst~xGrid, type = "l",xlab = "x",ylab = "density")
+#' abline(v = c(1,5),col = "gray")
+#' 
+#' # example 2: true parameters, which should be rediscovered
+#' theta0 <- c(mu = 1, sigma = 0.4)
+#' mle <- exp(theta0[1] - theta0[2]^2)
+#' perc <- 0.975		# some upper percentile, proxy for an upper bound
+#' upper <- qlnorm(perc, meanlog = theta0[1], sdlog = theta0[2])
+#' (thetaEst <- getParmsLognormForModeAndUpper(
+#'   mle,upper = upper,sigmaFac = qnorm(perc)) )
+#' 
+#' #plot the true and the rediscovered distributions
+#' xGrid = seq(0,10, length.out = 81)[-1]
+#' dx <- dlnorm(xGrid, meanlog = theta0[1], sdlog = theta0[2])
+#' dxEst <- dlnorm(xGrid, meanlog = thetaEst[1], sdlog = thetaEst[2])
+#' plot( dx~xGrid, type = "l")
+#' #plot( dx~xGrid, type = "n")
+#' #overplots the original, coincide
+#' lines( dxEst ~ xGrid, col = "red", lty = "dashed")
+#' 
+#' # example 3: explore varying the uncertainty (the upper quantile)
+#' x <- seq(0.01,1.2,by = 0.01)
+#' mle = 0.2
+#' dx <- sapply(mle*2:8,function(q99){
+#'   theta = getParmsLognormForModeAndUpper(mle,q99,qnorm(0.99))
+#'   #dx <- dDistr(x,theta[,"mu"],theta[,"sigma"],trans = "lognorm")
+#'   dx <- dlnorm(x,theta[,"mu"],theta[,"sigma"])
+#' })
+#'   matplot(x,dx,type = "l")
+
 getParmsLognormForModeAndUpper <- function(
   mle, upper,sigmaFac = qnorm(0.99)){
   # solution of
@@ -79,45 +117,6 @@ getParmsLognormForModeAndUpper <- function(
   sigma <- -f2 + sqrt(f2^2 + (log(upper) - m))
   mu <- m + sigma^2
   cbind(mu = mu, sigma = sigma )
-}
-example_getParmsLognormForModeAndUpper <- function(){
-  # example 1: a distribution with mode 1 and upper bound 5
-  (thetaEst <- getParmsLognormForModeAndUpper(1,5))
-  mle <- exp(thetaEst[1] - thetaEst[2]^2)
-  all.equal(mle, 1, check.attributes = FALSE)
-
-  # plot the distributions
-  xGrid = seq(0,8, length.out = 81)[-1]
-  dxEst <- dlnorm(xGrid, meanlog = thetaEst[1], sdlog = thetaEst[2])
-  plot( dxEst~xGrid, type = "l",xlab = "x",ylab = "density")
-  abline(v = c(1,5),col = "gray")
-
-  # example 2: true parameters, which should be rediscovered
-  theta0 <- c(mu = 1, sigma = 0.4)
-  mle <- exp(theta0[1] - theta0[2]^2)
-  perc <- 0.975		# some upper percentile, proxy for an upper bound
-  upper <- qlnorm(perc, meanlog = theta0[1], sdlog = theta0[2])
-  (thetaEst <- getParmsLognormForModeAndUpper(
-    mle,upper = upper,sigmaFac = qnorm(perc)) )
-
-  #plot the true and the rediscovered distributions
-  xGrid = seq(0,10, length.out = 81)[-1]
-  dx <- dlnorm(xGrid, meanlog = theta0[1], sdlog = theta0[2])
-  dxEst <- dlnorm(xGrid, meanlog = thetaEst[1], sdlog = thetaEst[2])
-  plot( dx~xGrid, type = "l")
-  #plot( dx~xGrid, type = "n")
-  #overplots the original, coincide
-  lines( dxEst ~ xGrid, col = "red", lty = "dashed")
-
-  # example 3: explore varying the uncertainty (the upper quantile)
-  x <- seq(0.01,1.2,by = 0.01)
-  mle = 0.2
-  dx <- sapply(mle*2:8,function(q99){
-    theta = getParmsLognormForModeAndUpper(mle,q99,qnorm(0.99))
-    #dx <- dDistr(x,theta[,"mu"],theta[,"sigma"],trans = "lognorm")
-    dx <- dlnorm(x,theta[,"mu"],theta[,"sigma"])
-  })
-  matplot(x,dx,type = "l")
 }
 
 #' @describeIn getParmsLognormForModeAndUpper 
@@ -146,20 +145,19 @@ getParmsLognormForMoments <- function(mean, var, sigmaOrig = sqrt(var)){
 #'   Calculate mu and sigma from expected value and geometric standard deviation
 #' @param median geometric mu (median at the original exponential scale)
 #' @param sigmaStar multiplicative standard deviation
-#' @exampleFunction example_getParmsLognormForExpval
+#' @examples
+#' # Calculate mu and sigma from expected value and geometric standard deviation
+#' .mean <- 1
+#' .sigmaStar <- c(1.3,2)
+#' (parms <- getParmsLognormForExpval(.mean, .sigmaStar))
+#' # multiplicative standard deviation must equal the specified value
+#' cbind(exp(parms[,"sigma"]), .sigmaStar)
+#' 
 #' @export
 getParmsLognormForExpval <- function(mean, sigmaStar){
   sigma = log(sigmaStar)
   mu = log(mean) - sigma*sigma/2
   cbind(mu = mu, sigma = sigma)
-}
-example_getParmsLognormForExpval <- function(){
-  # Calculate mu and sigma from expected value and geometric standard deviation
-  .mean <- 1
-  .sigmaStar <- c(1.3,2)
-  (parms <- getParmsLognormForExpval(.mean, .sigmaStar))
-  # multiplicative standard deviation must equal the specified value
-  cbind(exp(parms[,"sigma"]), .sigmaStar)
 }